Looking for Humanoid Object Fix
- reverted looking_for_humanoid_obj() back to algorithms.py
- Fixed references in _init_.py
