+++ algorithms.py
@@ -118,7 +118,7 @@ def get_bounding_box(v_coords):
 
     return None
 
-
+#TODO put this into file_ops
 def load_bbox_data(filepath):
     bboxes = []
     database_file = open(filepath, "r")
@@ -288,6 +288,42 @@ def correct_morph(base_form, current_form, morph_deltas, bboxes):
     logger.info("Morphing corrected in %s secs", time.time()-time1)
     return new_morph_deltas
 
+def looking_for_humanoid_obj():
+    """
+    Looking for a mesh that is OK for the lab
+    """
+    logger.info("Looking for a humanoid object ...")
+    if bpy.app.version < (2, 80, 74):
+        msg = "Sorry, MB-Lab requires Blender 2.80.74 Minimum"
+        logger.warning(msg)
+        return("ERROR", msg)
+
+#        if bpy.app.version >= (2,80,0):
+#            msg = "Sorry, this version of lab does no work with Blender 2.8"
+#            logger.warning(sg)
+#            return("ERROR",msg)
+
+#       if bpy.app.version > (2,79,0):
+#            msg = "The lab is not designed to work with unstable Blender build {0}".format(str(bpy.app.version))
+#            logger.warning(sg)
+        # return("ERROR",msg)
+
+    human_obj = None
+    name = ""
+    for obj in bpy.data.objects:
+        if obj.type == "MESH":
+            if "manuellab_vers" in get_object_keys(obj):
+                if check_version(obj["manuellab_vers"]):
+                    human_obj = obj
+                    name = human_obj.name
+                    break
+
+    if not human_obj:
+        msg = "No lab humanoids in the scene"
+        logger.info(msg)
+        return "NO_OBJ", msg
+
+    return "FOUND", name
 
 
 
@@ -378,7 +414,7 @@ def kdtree_from_mesh_vertices(mesh):
     research_tree.balance()
     return research_tree
 
-
+#TODO Put this in file_ops
 def import_mesh_from_lib(lib_filepath, name):
     existing_mesh_names = collect_existing_meshes()
     append_mesh_from_library(lib_filepath, [name])
